requirements.txt
google
google-cloud-vision
Pillow